:recycle: refactor : emitUserActivity 메소드 ActivityGateway 로 이동 #97
