+++ backend/src/user/dto/user-gateway.dto.ts
@@ -1,10 +1,4 @@
-import { Activity, Relationship, UserId } from '../../util/type';
-
-export interface UserActivityDto {
-  activity: Activity;
-  gameId: number;
-  userId: UserId;
-}
+import { Relationship, UserId } from '../../util/type';
 
 export interface UserRelationshipDto {
   relationship: Relationship | 'normal';
